+++ lib/__init__.py
@@ -1,2 +1,5 @@
-from .census_geocoder_api_client import CensusGeocoderApiClient, CensusGeocoderApiClientError # noqa
-from .nyc_geocoder_client import NycGeocoderClient # noqa
+from .census_geocoder_api_client import (
+    CensusGeocoderApiClient,
+    CensusGeocoderApiClientError,
+)
+from .nyc_geocoder_client import NycGeocoderClient